refactor: extract install logic to core/services, merge InstallationManager into Interface
